Add custom props to the `LoadingIndicator`
